Retrieve blocks command

+++ samples/retrieve/getblocks.js
@@ -15,6 +15,10 @@ const client = rskapi.client(config.host);
 const ndigits = to - from > 200000 ? 2 : 1;
 
 (async function() {
+    fs.mkdirSync(config.data, { recursive: true });
+    
+    const stream = fs.createWriteStream(path.join(config.data, 'chain.txt'), { flags: 'a' });
+    
     for (let k = from; k <= to; k++) {
         const block = await client.block(k, true);
         const hash = block.hash;
@@ -25,11 +29,18 @@ const ndigits = to - from > 200000 ? 2 : 1;
         
         const filename = path.join(dirname, hash + '.json');
         
+        if (fs.existsSync(filename))
+            continue;
+        
         fs.writeFileSync(filename, JSON.stringify(block));
+        
+        stream.write(k + ',' + hash + '\r\n');
         
         if (k % 100 == 0)
             process.stdout.write('.');
     }
+    
+    stream.end();
     
     console.log('done');
 })();
